+++ epochX/cudacpp/gg_ttg.sa/SubProcesses/EventStatistics.h
@@ -134,7 +134,7 @@ namespace mg5amcCpu
     void printout( std::ostream& out ) const
     {
       const EventStatistics& s = *this;
-      constexpr size_t meGeVexponent = -( 2 * CPPProcess::npar - 8 );
+      constexpr int meGeVexponent = -( 2 * CPPProcess::npar - 8 );
       out << s.tag << "NumMatrixElems(notAbnormal) = " << s.nevtOK() << std::endl
           << std::scientific // fixed format: affects all floats (default precision: 6)
           << s.tag << "MeanMatrixElemValue         = ( " << s.meanME()
