+++ epochX/cudacpp/gq_ttq.sa/SubProcesses/runTest.cc
@@ -24,8 +24,8 @@ using namespace mg5amcCpu;
 struct CUDA_CPU_TestBase : public TestDriverBase
 {
   static constexpr int neppM = MemoryAccessMomenta::neppM; // AOSOA layout
-  static constexpr size_t np4 = CPPProcess::np4;
-  static constexpr size_t npar = CPPProcess::npar;
+  static constexpr int np4 = CPPProcess::np4;
+  static constexpr int npar = CPPProcess::npar;
   static_assert( gputhreads % neppM == 0, "ERROR! #threads/block should be a multiple of neppM" );
   static_assert( gputhreads <= mgOnGpu::ntpbMAX, "ERROR! #threads/block should be <= ntpbMAX" );
   CUDA_CPU_TestBase( const std::string& refFileName )
@@ -100,7 +100,7 @@ struct CPUTest : public CUDA_CPU_TestBase
     mek.computeMatrixElements( channelId );
   }
 
-  fptype getMomentum( std::size_t ievt, size_t ipar, size_t ip4 ) const override
+  fptype getMomentum( std::size_t ievt, unsigned int ipar, unsigned int ip4 ) const override
   {
     assert( ipar < npar );
     assert( ip4 < np4 );
@@ -219,7 +219,7 @@ struct CUDATest : public CUDA_CPU_TestBase
     copyHostFromDevice( hstMatrixElements, devMatrixElements );
   }
 
-  fptype getMomentum( std::size_t ievt, size_t ipar, size_t ip4 ) const override
+  fptype getMomentum( std::size_t ievt, unsigned int ipar, unsigned int ip4 ) const override
   {
     assert( ipar < npar );
     assert( ip4 < np4 );
